# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplikasi Gacor API Documentation",
        "description": "Documentation for the Aplikasi Gacor Package Delivery Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/pengirims": {
            "get": {
                "tags": [
                    "Pengirim"
                ],
                "summary": "Ambil semua data pengirim",
                "operationId": "ea54a0d95cfff1aac0b6d4475c2e6688",
                "responses": {
                    "200": {
                        "description": "Data pengirim berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengirim"
                ],
                "summary": "Buat pengirim baru",
                "operationId": "b74bc851bc13936673935eae23b8cfaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "alamat",
                                    "nomor_telepon"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "PT Express"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Gatot Subroto No. 789, Jakarta"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "021-555-0123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengirim berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pengirims/{id}": {
            "get": {
                "tags": [
                    "Pengirim"
                ],
                "summary": "Ambil detail pengirim",
                "operationId": "ca67deb5be8ea169cb4a5cfa86a1e4d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengirim ditemukan"
                    },
                    "404": {
                        "description": "Pengirim tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pengirim"
                ],
                "summary": "Update data pengirim",
                "operationId": "c1ec967dcb7b54e02b73bcd04d6f89fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "PT Delivery Plus"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. HR Rasuna Said No. 456, Jakarta"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "021-555-0124"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengirim berhasil diperbarui"
                    },
                    "404": {
                        "description": "Pengirim tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pengirim"
                ],
                "summary": "Hapus pengirim",
                "operationId": "b2ace8422b931065d2db35666835067f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengirim berhasil dihapus"
                    },
                    "404": {
                        "description": "Pengirim tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pengirims/filter": {
            "get": {
                "tags": [
                    "Pengirim"
                ],
                "summary": "Filter data pengirim berdasarkan nama atau nomor telepon",
                "operationId": "78df7bd477db8b38db1696a42fcee846",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Filter berdasarkan nama",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nomor_telepon",
                        "in": "query",
                        "description": "Filter berdasarkan nomor telepon",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pengirim hasil filter"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}